+++ R/zzz.R
@@ -20,6 +20,7 @@ default_linters <- modify_defaults(
   cyclocomp_linter(),
   equals_na_linter(),
   function_left_parentheses_linter(),
+  indentation_linter(),
   infix_spaces_linter(),
   line_length_linter(),
   no_tab_linter(),